% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate-implied-volatility-bid-ask.R
\name{calculate_implied_volatility}
\alias{calculate_implied_volatility}
\title{Title Implied volatility with American option DB}
\usage{
calculate_implied_volatility(option_data, option_distributions,
  option_risk_free_rates, security_data, save_path, save_filename, cl)
}
\arguments{
\item{option_data}{}

\item{option_distributions}{}

\item{option_risk_free_rates}{}

\item{security_data}{}

\item{save_path}{}

\item{save_filename}{}

\item{cl}{}
}
\description{
Title Implied volatility with American option DB
}
