% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omtr-fetch-equity-options-and-underlying-data.R
\name{fetch_equity_option_data}
\alias{fetch_equity_option_data}
\title{Title extract equity option data with associated distributions and equity prices}
\usage{
fetch_equity_option_data(secids, save_path)
}
\arguments{
\item{secids}{}

\item{save_path}{}
}
\description{
Title extract equity option data with associated distributions and equity prices
}
