% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omtr-fetch-option-volume-data.R
\name{fetch_option_volume_data}
\alias{fetch_option_volume_data}
\title{Title extract option volume data}
\usage{
fetch_option_volume_data(secids, save_path)
}
\arguments{
\item{secids}{}

\item{save_path}{}
}
\description{
Title extract option volume data
}
