% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crsp-fetch-equity-returns.R
\name{fetch_stock_data}
\alias{fetch_stock_data}
\title{CRSP stock data}
\usage{
fetch_stock_data(permnos, frequency = "daily", date_start = NULL,
  date_end = NULL, save_path = NULL)
}
\arguments{
\item{permnos}{}

\item{frequency}{}

\item{date_start}{}

\item{date_end}{}

\item{save_path}{}
}
\description{
CRSP stock data
}
