% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate-and-match-zero-coupon-curve.R
\name{interpolate_zero_curve_for_options}
\alias{interpolate_zero_curve_for_options}
\title{Zero curve interpolation for option panel}
\usage{
interpolate_zero_curve_for_options(option_data, zero_coupon_data)
}
\arguments{
\item{option_data}{}

\item{zero_coupon_data}{}
}
\description{
Zero curve interpolation for option panel
}
