% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match-options-and-dividends-calc-pv.R
\name{match_distributions_with_options}
\alias{match_distributions_with_options}
\title{Matching of options and distributions, pv of distributions}
\usage{
match_distributions_with_options(option_data, distributions_data,
  zero_coupon_data)
}
\arguments{
\item{option_data}{}

\item{zero_coupon_data}{}
}
\description{
Matching of options and distributions, pv of distributions
}
